classifier uses new serialization

+++ python/baseline/remote.py
@@ -91,7 +91,7 @@ class RemoteModelTensorFlowREST(object):
             for i in range(num_ex):
                 score_i = scores[i]
                 classes_i = classes[i]
-                d = [(c, s) for c, s in zip(classes_i, score_i)]
+                d = [(c, np.float32(s)) for c, s in zip(classes_i, score_i)]
                 result.append(d)
             return result
